BREAKING CHANGE: Remove level-sublevel dependency
Problem: This dependency is unmaintained and not really necessary. It
allows us to embed multiple Level database instances in one, but we
don't actually need that for any reason. This module is the source of an
`npm audit` warning, and also may be the source of a new bug that's been
happening in Oasis.

Solution: Don't do that, and let level have its own database. The only
problem is that **any pending invites will be ignored** because this
commit doesn't include a migration script. This is fine for users to
upgrade to, but pubs that still have lots of pending invites (?) may
want to wait until their invites are redeemed (or write a migration
script).

Fixes: https://github.com/ssbc/ssb-invite/issues/13